test: refactor chdir mocha test to tap
